[build] rename example platform POSIX to sim
People are confused about the two POSIX platforms: example POSIX and
POSIX-app.  This PR renames the example platform POSIX to
sim(simulation), and after this is merge, another PR will rename the
current POSIX-app to POSIX.
