+++ .travis.yml
@@ -81,23 +81,23 @@ jobs:
     - env: BUILD_TARGET="gn-build" VERBOSE=1
       os: linux
       script: .travis/script.sh
-    - env: BUILD_TARGET="posix-distcheck" VERBOSE=1 VIRTUAL_TIME=1
+    - env: BUILD_TARGET="sim-distcheck" VERBOSE=1 VIRTUAL_TIME=1
       os: linux
       compiler: clang
       script: .travis/script.sh
-    - env: BUILD_TARGET="posix-32-bit" VERBOSE=1 VIRTUAL_TIME=1
+    - env: BUILD_TARGET="sim-32-bit" VERBOSE=1 VIRTUAL_TIME=1
       os: linux
       compiler: gcc
       script: .travis/script.sh
-    - env: BUILD_TARGET="posix-32-bit" VERBOSE=1 VIRTUAL_TIME=1 TIME_SYNC=1
+    - env: BUILD_TARGET="sim-32-bit" VERBOSE=1 VIRTUAL_TIME=1 TIME_SYNC=1
       os: linux
       compiler: gcc
       script: .travis/script.sh
-    - env: BUILD_TARGET="posix-ncp" VERBOSE=1 VIRTUAL_TIME=1
+    - env: BUILD_TARGET="sim-ncp" VERBOSE=1 VIRTUAL_TIME=1
       os: linux
       compiler: gcc
       script: .travis/script.sh
-    - env: BUILD_TARGET="posix-mtd" VERBOSE=1 VIRTUAL_TIME=1
+    - env: BUILD_TARGET="sim-mtd" VERBOSE=1 VIRTUAL_TIME=1
       os: linux
       compiler: gcc
       script: .travis/script.sh
@@ -131,7 +131,7 @@ jobs:
       os: linux
       compiler: gcc
       script: .travis/script.sh
-    - env: BUILD_TARGET="posix" CC="gcc-5" CXX="g++-5"
+    - env: BUILD_TARGET="sim" CC="gcc-5" CXX="g++-5"
       os: linux
       compiler: gcc
       addons:
@@ -140,7 +140,7 @@ jobs:
             - gcc-5
             - g++-5
       script: .travis/script.sh
-    - env: BUILD_TARGET="posix" CC="gcc-6" CXX="g++-6"
+    - env: BUILD_TARGET="sim" CC="gcc-6" CXX="g++-6"
       os: linux
       compiler: gcc
       addons:
@@ -149,7 +149,7 @@ jobs:
             - gcc-6
             - g++-6
       script: .travis/script.sh
-    - env: BUILD_TARGET="posix" CC="gcc-7" CXX="g++-7"
+    - env: BUILD_TARGET="sim" CC="gcc-7" CXX="g++-7"
       os: linux
       compiler: gcc
       addons:
@@ -158,7 +158,7 @@ jobs:
             - gcc-7
             - g++-7
       script: .travis/script.sh
-    - env: BUILD_TARGET="posix" CC="gcc-8" CXX="g++-8"
+    - env: BUILD_TARGET="sim" CC="gcc-8" CXX="g++-8"
       os: linux
       compiler: gcc
       addons:
@@ -167,7 +167,7 @@ jobs:
             - gcc-8
             - g++-8
       script: .travis/script.sh
-    - env: BUILD_TARGET="posix" CC="gcc-9" CXX="g++-9"
+    - env: BUILD_TARGET="sim" CC="gcc-9" CXX="g++-9"
       os: linux
       compiler: gcc
       addons:
@@ -178,7 +178,7 @@ jobs:
             - gcc-9
             - g++-9
       script: .travis/script.sh
-    - env: BUILD_TARGET="posix-ncp-spi" VERBOSE=1
+    - env: BUILD_TARGET="sim-ncp-spi" VERBOSE=1
       os: linux
       compiler: gcc
       script: .travis/script.sh
