+++ examples/platforms/sim/flash.c
@@ -26,7 +26,7 @@
  *  POSSIBILITY OF SUCH DAMAGE.
  */
 
-#include "platform-posix.h"
+#include "platform-sim.h"
 
 #include <fcntl.h>
 #include <stdio.h>
