+++ examples/platforms/sim/platform-sim.h
